// src/lib/mail/types.ts
export type DefaultSmtpConfig = SMTPConnection.Options;

/**
 * 邮件客户端基础配置
 */
interface BaseMailClientConfig {
    /**
     * 配置名称
     */
    name: string;
}
// ...
interface SmtpMailClientConfig extends BaseMailClientConfig {
    type: "smtp";
    options: DefaultSmtpConfig;
    default: Required<Pick<BaseMailSendOptions, "from">> & Pick<BaseMailSendOptions, "reply">;
}
// ...
export interface SmtpMailClient extends Omit<SmtpMailClientConfig, "options"> {
    client: Transporter<SMTPTransport.SentMessageInfo, SMTPTransport.Options>;
}

/**
 * 邮件客户端类型
 */
export type MailClient = AliyunMailClient | TencentMailClient | SmtpMailClient;

/**
 * 邮件发送函数选项参数基础类型
 */
export interface BaseMailSendOptions {
    /**
     * 发件人邮箱地址
     */
    from?: string;
    /**
     * 发件人名称
     */
    fromName?: string;
    /**
     * 收件人邮箱地址列表
     */
    to: string[];
    /**
     * 邮件主题
     */
    subject: string;
    /**
     * 抄送人邮箱地址列表
     */
    cc?: string[];
    /**
     * 密送人邮箱地址列表
     */
    bcc?: string[];
    /**
     * 回复邮箱地址
     */
    reply?: string;
    /**
     * 回复邮箱名称
     */
    replyName?: string;
    /**
     * 变量
     */
    vars: Record<string, any>;
}
// ...
export type SmtpSendMailOptions = Omit<BaseMailSendOptions, "subject"> & { subject?: string } & {
    templatePath: string;
    others?: Omit<Mail.Options, "replyTo" | "replyName">;
};